+++ idea/src/org/jetbrains/kotlin/idea/refactoring/introduce/extractClass/ui/KotlinExtractSuperDialogBase.kt
@@ -71,6 +71,13 @@ abstract class KotlinExtractSuperDialogBase(
             return !declaration.hasModifier(KtTokens.CONST_KEYWORD)
         }
 
+        override fun isAbstractEnabled(memberInfo: KotlinMemberInfo): Boolean {
+            val member = memberInfo.member
+            return !(member.hasModifier(KtTokens.INLINE_KEYWORD) ||
+                     member.hasModifier(KtTokens.EXTERNAL_KEYWORD) ||
+                     member.hasModifier(KtTokens.LATEINIT_KEYWORD))
+        }
+
         override fun isFixedAbstract(memberInfo: KotlinMemberInfo?) = true
     }