+++ idea/src/org/jetbrains/kotlin/idea/refactoring/introduce/extractClass/ui/KotlinExtractSuperclassDialog.kt
@@ -51,6 +51,7 @@ class KotlinExtractSuperclassDialog(
                 getInterfaceContainmentVerifier { selectedMembers }
         ) {
             override fun isAbstractEnabled(memberInfo: KotlinMemberInfo): Boolean {
+                if (!super.isAbstractEnabled(memberInfo)) return false
                 val member = memberInfo.member
                 return member is KtNamedFunction || member is KtProperty || member is KtParameter
             }